Update README.md
Resolved an issue with the Quick Tips sections - it was unfinished. We are now discussing some important tips on how a student would best learn Swift and building apps on iOS. We suggest building tutorials as brain training best practice to learn Swift and also to retain the information also. Writing tutorials to teach others is an important tool in the learning process as outlined by world renowned learning expert Jim Kwik!